Griffon Greed
Added in the modifiers to incentivize griffons to be greedy.  This was
done via the addition of an ai_greed line to the griffon species trait,
and six triggered modifiers to give them more prestige and greed the
more gold they have.  Also added the modifier for the idol of boreas,
still need to make the events to give that modifier out.  (base it on
the events that give holy relics in vanilla?)